云信通话集成，除了登录接口调试，其他都完成

+++ NoticeXi/Common/NoticeAudioChatTools.h
@@ -33,7 +33,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic,assign) BOOL noReClick;//防止重复点击
 
-- (void)callToUserId:(NSString *)userId roomId:(UInt32)roomIdNum getOrderTime:(NSString *)getOrderTime nickName:(NSString *)nickName autoNext:(BOOL)autonext;
+- (void)callToUserId:(NSString *)userId roomId:(NSInteger)roomIdNum getOrderTime:(NSString *)getOrderTime nickName:(NSString *)nickName autoNext:(BOOL)autonext;
 @property (nonatomic, strong) NoticeCallView * __nullable callingView;
 @end
